Move CIS2 methods to their own module
